wip testDelayedJobExecutionThrowsWhenConnectionFails
